requirements.txt
Django==4.0.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
django-cors-headers==3.11.0
django-linear-migrations==2.3.0